Move sequence-number utils to ens folder aswell
